Make Husky a dev dependency
When Husky is installed, it takes control of the Git hooks in whatever repository the install is happening in the `node_modules` of. So right now, if any project depends on `@truffle/contract@4.2.9`, it gets its git hooks modified by Husky, because Husky is eventually a transitive non-dev dependency via this project.

This is unacceptable and highly suspicious behavior from a package way down in the dependency tree.

This commit makes Husky a dev dependency, so it should only act on the Git repo of this project that actually wants it.